requirements.txt
torch==2.0.1 --index-url https://download.pytorch.org/whl/cu118
accelerate==0.25.0
auto-gptq==0.4.2
datasets==2.14.6
deepspeed==0.11.1
evaluate==0.4.0
faiss-gpu
fastapi
fire==0.5.0
flash-attn==2.3.3
gradio
huggingface-hub==0.20.1
jieba==0.42.1
jsonlines
matplotlib
nltk==3.8.1
numpy
nvitop
openai
-e git+https://github.com/LZY-the-boys/peft@953a014fb76211cfd5e2c9f0d1497731e101d4ad#egg=peft
pydantic==1.10.13
pydantic_core==2.10.1
rouge
rouge-score
safetensors==0.4.0
scikit-learn
scipy==1.5.4
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
tokenizers==0.15.0
torchaudio==2.0.2
torchmetrics==0.9.0
torchvision==0.15.2
tqdm
transformers==4.36.2
uvicorn==0.23.2
tabulate
defopt
openpyxl
einops 
transformers_stream_generator
tiktoken
fastapi
wandb